Add port collapsing
