+++ app/scripts/react-directives/device-manager/pageStore.js
@@ -1,13 +1,21 @@
 'use strict';
 
-import { makeAutoObservable, makeObservable, observable, computed, action } from 'mobx';
-import { cloneDeep, isEqual } from 'lodash';
+import { makeObservable, observable, computed } from 'mobx';
+import { cloneDeep } from 'lodash';
 import i18n from '../../i18n/react/config';
 import SelectModalState from '../components/modals/selectModalState';
 import { getDefaultObject } from './jsonSchemaUtils';
 import AccessLevelStore from '../components/access-level/accessLevelStore';
 import PageWrapperStore from '../components/page-wrapper/pageWrapperStore';
 import ConfirmModalState from '../components/modals/confirmModalState';
+import {
+  Tab,
+  TabsStore,
+  makeDeviceTabName,
+  makeModbusTcpPortTabName,
+  makeSerialPortTabName,
+  makeTcpPortTabName,
+} from './tabsStore';
 
 function getTranslation(key, translations) {
   return translations[i18n.language]?.[key] || translations?.en?.[key] || key;
@@ -84,22 +92,6 @@ function getPortData(data) {
   return res;
 }
 
-function makeSerialPortTabName(data, schema) {
-  return data?.path?.replace(/^\/dev\/tty/, '');
-}
-
-function makeTcpPortTabName(data, schema) {
-  return `TCP ${data.address || ''}:${data.port || ''}`;
-}
-
-function makeModbusTcpPortTabName(data, schema) {
-  return `MODBUS TCP ${data.address || ''}:${data.port || ''}`;
-}
-
-function makeDeviceTabName(data, schema) {
-  return `${data.slave_id || ''} ` + getTranslation(schema.title, schema.translations);
-}
-
 function makeDeviceTypeSelectOptions(deviceSchemaMap) {
   let groups = {};
   let wbDevicesGroupName;
@@ -149,134 +141,6 @@ function makeDeviceTypeSelectOptions(deviceSchemaMap) {
       }
       return a.label.localeCompare(b.label);
     });
-}
-
-class Tab {
-  constructor(type, data, schema, nameGenerationFn) {
-    this.name = '';
-    this.type = type;
-    this.data = data;
-    this.editedData = cloneDeep(data);
-    this.schema = schema;
-    this.isValid = true;
-    this.isDirty = false;
-    this.nameGenerationFn = nameGenerationFn;
-
-    this.updateName();
-
-    makeObservable(this, {
-      name: observable,
-      isValid: observable,
-      isDirty: observable,
-      setData: action.bound,
-      updateName: action,
-      submit: action,
-    });
-  }
-
-  updateName() {
-    this.name = this.nameGenerationFn(this.editedData, this.schema);
-  }
-
-  setData(data, errors) {
-    this.isDirty = !isEqual(this.data, data);
-    this.editedData = cloneDeep(data);
-    this.isValid = errors.length == 0;
-    this.updateName();
-  }
-
-  submit() {
-    this.data = cloneDeep(this.editedData);
-    this.isValid = true;
-    this.isDirty = false;
-  }
-}
-
-class TabsStore {
-  constructor() {
-    this.items = [];
-    this.selectedTabIndex = 0;
-    this.hasNewOrDeletedItems = false;
-
-    makeAutoObservable(this);
-  }
-
-  addTab(tab, initial) {
-    if (tab.type == 'port') {
-      this.items.push(tab);
-      if (!initial) {
-        this.selectedTabIndex = this.items.length - 1;
-        this.hasNewOrDeletedItems = true;
-      }
-      return;
-    }
-    let i = this.selectedTabIndex + 1;
-    while (i < this.items.length && this.items[i]?.type != 'port') {
-      i++;
-    }
-    this.items.splice(i, 0, tab);
-    if (!initial) {
-      this.selectedTabIndex = i;
-      this.hasNewOrDeletedItems = true;
-    }
-  }
-
-  onSelectTab(index, lastIndex) {
-    this.selectedTabIndex = index;
-    return true;
-  }
-
-  selectTab(index) {
-    this.selectedTabIndex = index;
-  }
-
-  deleteSelectedTab() {
-    let count = 1;
-    if (this.items[this.selectedTabIndex]?.type == 'port') {
-      while (
-        this.selectedTabIndex + count < this.items.length &&
-        this.items[this.selectedTabIndex + count]?.type == 'device'
-      ) {
-        count++;
-      }
-    }
-    this.items.splice(this.selectedTabIndex, count);
-    this.selectedTabIndex = 0;
-    this.hasNewOrDeletedItems = true;
-  }
-
-  get selectedPortTab() {
-    let i = this.selectedTabIndex;
-    while (i >= 0 && this.items[i]?.type != 'port') {
-      i--;
-    }
-    if (i >= this.items.length) {
-      return undefined;
-    }
-    return this.items[i];
-  }
-
-  get isValid() {
-    return this.items.every(item => item.isValid);
-  }
-
-  get isDirty() {
-    return this.hasNewOrDeletedItems || this.items.some(item => item.isDirty);
-  }
-
-  get isEmpty() {
-    return this.items.length == 0;
-  }
-
-  submit() {
-    this.items.forEach(item => item.submit());
-    this.hasNewOrDeletedItems = false;
-  }
-
-  clear() {
-    this.items.splice(0, this.items.length);
-    this.hasNewOrDeletedItems = false;
-  }
 }
 
 class DeviceManagerPageStore {
@@ -462,7 +326,7 @@ class DeviceManagerPageStore {
     this.pageWrapperStore.clearError();
     try {
       await this.saveConfigFn(this.makeConfigJson());
-      this.tabs.submit();
+      this.tabs.commitData();
     } catch (err) {
       if (err.data === 'EditorError' && err.code === CONFED_WRITE_FILE_ERROR) {
         this.pageWrapperStore.setError(i18n.t('device-manager.errors.write'));
